+++ src/Servers/OSX.php
@@ -19,7 +19,6 @@ use Innmind\Server\Control\Server as Control;
 use Innmind\TimeContinuum\Clock;
 use Innmind\Url\Path;
 use Innmind\Immutable\Attempt;
-use Innmind\Immutable\Maybe;
 
 final class OSX implements Server
 {
@@ -45,7 +44,7 @@ final class OSX implements Server
     }
 
     #[\Override]
-    public function memory(): Maybe
+    public function memory(): Attempt
     {
         return ($this->memory)();
     }